Partial port of status/EventsFilter.ts
